# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Picks League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/events/{eventId}/my-submission:
    get:
      operationId: getMySubmission
      tags: [submissions]
      summary: Get current user's submission for an event
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Submission (null if not yet submitted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /leagues/{leagueId}/events/{eventId}/submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions (after reveal) or submission count (before reveal)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Submissions or count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: submitPicks
      tags: [submissions]
      summary: Submit picks for an event
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /leagues/{leagueId}/events/{eventId}/submissions/{submissionId}:
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update picks before deadline
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
        - name: submissionId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Live Board ───────────────────────────────────────────────────────────────
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: number
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: number
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Submission:
      type: object
      required: [id, pickEventId, userId, submittedAt, picks]
      properties:
        id:
          type: number
        pickEventId:
          type: number
        userId:
          type: string
        tiebreakerAnswer:
          type: ["number", "null"]
        moneyPickGameId:
          type: ["number", "null"]
        submittedAt:
          type: string
          format: date-time
        lockedAt:
          type: ["string", "null"]
          format: date-time
        picks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
    SubmissionEnvelope:
      type: object
      required: [submission]
      properties:
        submission:
          oneOf:
            - $ref: '#/components/schemas/Submission'
            - type: 'null'
    SubmissionsResponse:
      type: object
      required: [revealed, submittedCount, totalMembers]
      properties:
        revealed:
          type: boolean
        submittedCount:
          type: number
        totalMembers:
          type: number
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionWithUser"
    SubmissionWithUser:
      type: object
      required: [id, pickEventId, userId, submittedAt, picks, displayName, profileImageUrl]
      properties:
        id:
          type: number
        pickEventId:
          type: number
        userId:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        tiebreakerAnswer:
          type: ["number", "null"]
        moneyPickGameId:
          type: ["number", "null"]
        submittedAt:
          type: string
          format: date-time
        lockedAt:
          type: ["string", "null"]
          format: date-time
        picks:
          type: array
          items:
            $ref: "#/components/schemas/Pick"
    Pick:
      type: object
      required: [id, submissionId, eventGameId, selectedTeam]
      properties:
        id:
          type: number
        submissionId:
          type: number
        eventGameId:
          type: number
        selectedTeam:
          type: string
          enum: [home, away]
        result:
          type: ["string", "null"]
          enum: ["win", "loss", "push", null]
        pointsAwarded:
          type: ["number", "null"]
    SubmissionInput:
      type: object
      required: [picks, moneyPickGameId, tiebreakerAnswer]
      properties:
        picks:
          type: array
          items:
            $ref: "#/components/schemas/PickInput"
        moneyPickGameId:
          type: number
        tiebreakerAnswer:
          type: number
    PickInput:
      type: object
      required: [eventGameId, selectedTeam]
      properties:
        eventGameId:
          type: number
        selectedTeam:
          type: string
          enum: [home, away]
    SubmissionUpdate:
      type: object
      required: [picks, moneyPickGameId, tiebreakerAnswer]
      properties:
        picks:
          type: array
          items:
            $ref: "#/components/schemas/PickInput"
        moneyPickGameId:
          type: number
        tiebreakerAnswer:
          type: number

    # ── Live Board ────────────────────────────────────────────────────────────
...